New palette selection mechanism
